+++ backend/domain/payment-domain/src/main/java/org/atlas/domain/payment/entity/PaymentEntity.java
@@ -15,10 +15,10 @@ import org.atlas.framework.domain.entity.DomainEntity;
 public class PaymentEntity extends DomainEntity {
 
   @EqualsAndHashCode.Include
-  private Long id;
-  private Long sagaId;
-  private Long userId;
-  private Long orderId;
+  private Integer id;
+  private Integer sagaId;
+  private Integer userId;
+  private Integer orderId;
   private BigDecimal amount;
   private String currency;
   private PaymentMethod method;
