+++ backend/framework/src/main/java/org/atlas/framework/saga/context/SagaContext.java
@@ -11,7 +11,7 @@ public class SagaContext {
 
   @Getter
   @Setter
-  private Long sagaId;
+  private Integer sagaId;
 
   private final Map<String, Object> data = new ConcurrentHashMap<>();
