+++ backend/infrastructure/api-server/rest/api-server-rest-impl-order/src/main/java/org/atlas/infrastructure/api/server/rest/impl/order/controller/OrderController.java
@@ -98,7 +98,7 @@ public class OrderController {
     FrontCheckoutInput input = ObjectMapperUtil.getInstance()
         .map(request, FrontCheckoutInput.class);
 
-    Long sagaId = frontCheckoutUseCaseHandler.handle(input);
+    Integer sagaId = frontCheckoutUseCaseHandler.handle(input);
 
     CheckoutResponse response = CheckoutResponse.builder()
         .sagaId(sagaId)
