+++ backend/infrastructure/api-server/rest/api-server-rest-impl-order/src/main/java/org/atlas/infrastructure/api/server/rest/impl/order/model/CheckoutResponse.java
@@ -16,5 +16,5 @@ import lombok.Setter;
 public class CheckoutResponse {
 
   @Schema(description = "The identifier of saga")
-  private Long sagaId;
+  private Integer sagaId;
 }
